Added jsp for the generators

+++ homework4/cookie-clicker-lab-7-2/src/main/java/edu/csula/web/AdminGeneratorsRemoveServlet.java
@@ -12,10 +12,12 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
 import edu.csula.storage.GeneratorsDAO;
+import edu.csula.storage.UsersDAO;
 import edu.csula.storage.servlet.EventsDAOImpl;
 import edu.csula.storage.EventsDAO;
 import edu.csula.models.Event;
 import edu.csula.storage.servlet.GeneratorsDAOImpl;
+import edu.csula.storage.servlet.UsersDAOImpl;
 
 @WebServlet("/generatorsremove")
 
@@ -24,9 +26,15 @@ public class AdminGeneratorsRemoveServlet extends HttpServlet {
     @Override
     public void doGet( HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
         // TODO: handle upsert transaction
-        GeneratorsDAO dao = new GeneratorsDAOImpl(getServletContext());
-        int id=Integer.parseInt(request.getParameter("id"));
-        dao.remove(id);
-        response.sendRedirect("generators");
+        UsersDAO dao1=new UsersDAOImpl(request.getSession());
+        if(dao1 != null) {
+            GeneratorsDAO dao = new GeneratorsDAOImpl(getServletContext());
+            int id = Integer.parseInt(request.getParameter("id"));
+            dao.remove(id);
+            response.sendRedirect("generators");
+        }
+        else{
+            dao1.logout();
+        }
     }
 }
